deal with potential protocol_observer existence in the postgres plugin, graphs for latency coming from protocol_observer

+++ src/util/lib/cosi/plugins/index.js
@@ -87,6 +87,10 @@ class Plugin extends Events {
         });
     }
 
+    addCustomMetrics(regSetup) {
+        /* noop in base class */
+    }
+
     reregisterHost(quiet) {
         const self = this;
         this.once("reg-setup", () => {
@@ -94,6 +98,10 @@ class Plugin extends Events {
 
             regSetup.once("setup.done", () => {
                 self.emit("reg-config");
+            });
+
+            regSetup.once("metrics.fetch.done", () => {
+                self.addCustomMetrics(regSetup);
             });
 
             regSetup.setup();
@@ -121,8 +129,10 @@ class Plugin extends Events {
             regRegister.register();
         });
 
-        // metrics likely changed, delete cache of metrics to force a reget
-        fs.unlinkSync(path.resolve(this.regDir, "setup-metrics.json"));
+        if (fs.existsSync(path.resolve(this.regDir, "setup-metrics.json"))) {
+            // metrics likely changed, delete cache of metrics to force a reget
+            fs.unlinkSync(path.resolve(this.regDir, "setup-metrics.json"));
+        }
         
         // nad has a delay to pick up the newly linked metrics as the file system
         // watch takes a bit to trigger and then nad needs a moment to refresh 
@@ -132,7 +142,7 @@ class Plugin extends Events {
         }, 5000);
     }
 
-    configDashboard(name, quiet, dashboard_items) {
+    configDashboard(name, quiet, dashboard_items, fsGraphId) {
         const self = this;
         const regConfig = new RegConfig(quiet);
         const regSetup = new RegSetup(quiet);
@@ -146,7 +156,7 @@ class Plugin extends Events {
         });
 
         regSetup.once("templates.fetch.done", () => {
-            regConfig.configDashboard(name, dashboard_items);
+            regConfig.configDashboard(name, dashboard_items, fsGraphId);
         });        
 
         regSetup.fetchTemplates([`dashboard-${name}`]);
